pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "simulation"
version = "0.0.1"
description = "irp-yz11721 project"
requires-python = ">=3.8"

authors = [
    {name = "hahaha"}
]

classifiers = [
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3 :: Only"
]

dependencies = [
    "ubermagutil==0.63.0",
    "discretisedfield==0.64.0",
    "ubermagtable==0.61.1",
    "micromagneticmodel==0.63.0",
    "micromagneticdata==0.64.0",
    "micromagnetictests==0.62.0",
    "mumax3c==0.1",
    "oommfc==0.63.0",
    "mag2exp==0.62.0",
    "imageio==2.9.0",
    "ipython==8.5.0"
]

[project.optional-dependencies]
dev = [
    "build",
    "invoke",
    "nbval",
    "pre-commit",
    "pytest-cov",
    "twine"
]


[tool.black]
experimental-string-processing = true

[tool.coverage.run]
omit = ["tests/*"]

[tool.isort]
profile = "black"
skip_gitignore = true  # ignores files listed in .gitignore


[tool.pytest.ini_options]
pythonpath = [
    "continuous_model",
]
testpaths = [
    "tests",
]